feat: redesigned the js code.
Probably a number of commits but I have redesigned the layout of the code and how it was working.
Removed most of the functions and am working to try and make it work first and will then refactor into functions.
at the moment all the logic is inside the event listener function.